I have left broken HOD_Mock broken. Getting this module running with all the ho models is the priority.
Important experiment: creating a very large (Npts = 1e7)
numpy structured array is only marginally more expensive (2.e-5 seconds)
than creating a naked ndarray (1.e-5 seconds).
So I think the mock object should have a galaxies array
attached to it after all. If this can be implemented, it will dramatically simplify future
bookkeeping, such as mock shuffling routines. Needs to be played around with
to see if this is actually possible.

Note that the same is not true for an astropy table, which takes about 1000 times slower
to instantiate (0.02 seconds).