Continuing to develop make_mocks to be able to create an assembly-biased mock.
I have been going through and using the newly written halo_type_calculator
to assign halo types to host halos, in the event that a secondary halo property
key is not None.

Now the API is such that
even for simple Zheng07 type models, mean_ncen and mean_nsat and mean_concentration must
be called with primary_halo_property and halo_type input arrays. This is old news, but
I am still finding places where the baseline hod model tries to call one of its functions
with only primary_halo_property.

In general, I am still overhauling the way that the mock object accesses the halo catalog
and galaxy catalog. I have settled on an astropy table for mock.halos, but
I'm pretty sure I need to create a structured ndarray for the galaxies. This is now the point of focus.