assembias_HOD_quenching_model can now be used by make_mocks. However...
there is some weird behavior that I need to get to the bottom of.
When I tried to create a mock with assembias_quenching_parameters = [0.5,100],
the polynomial model solver was unable to find a solution within the required
number of iterations. First off, this is weird. Secondly, I should probably
just do away with this polynomial solver, and use an interpolater.