Changing interface between mock_factory and the composite HOD model.
Now, for the MC methods of the composite model, mock_factory will pass
the composite model the entire collection of halos. So now the composite model
will need to know what table columns to use when calling the component models.

Doing things this way simplifies the calls made to the model by the mock_factory,
and relegates the abstractions to hod_factory, where it belongs.