Now have working skeleton of Leauthaud11-style HOD. Only partially finished.
Still have lots of work to do on parameter passing/updating. Satellite model
is really just a skeleton. To complete this style model, it will also be necessary
to complete the SMHM model.

Now is probably the time to go ahead and finalize my conventions for kwargs
to model components.