Starting to implement significant changes to how models communicate with each other. Will get worse before it gets better.
It is no longer permissible to freely specify 'cens' or 'centrals'.
This freedom is quite pointless, and introduces needless complications.

If a model has ancillary dependencies, then its param_dict should reflect that.
The only problem with this is that the composite model will need to be notified in some way
not to complain when it detects multiple keys of the same name. I'm doing this
via the 'ancillary_model_param_dict' mechanism. Still not sure how this is going to work out.

Component models with ancillary dependence are now in charge of updating their
ancillary model's dictionary before each function call to the ancillary model.