Trivial change to readme for purposes of writing this lengthy commit message.
The branch hearin15_devel is where I have been developing the Hearin15 class
that governs the composite Decorated HOD model, which is based on Leauthaud11.
I noticed that I was not getting reasonable results for how assembly bias
scaled with the stellar mass threshold. In particular, I saw no weakening of
assembly bias as the stellar mass threshold increased, which cannot be correct.
So then I checked the satellite fractions of the Hearin15 mocks, and noticed
that no matter the threshold the fraction is about 30%, including for a
threshold of 11.5, for which the fraction should be about 15%. The same problem
occurs for the Leauthaud11 model, so I started a new branch hearin15_devel_fsat_debug,
and went to work on Leauthaud11.

After fiddling around with it for a while, with nothing in particular jumping out at me,
I decided that the only real way to reliably track the bug was to produce
Alexie's actual model, not just a crude approximation based on Moster13. So this meant
I had to go ahead and code up Peter Behroozi's 2010 SmHm model, which I did.
I also took away the freedom from Leauthaud11 to arbitrarily define a SmHm model -
now smhm_model is not a keyword option - Behroozi10 is automatically chosen.
Might relax this later, we'll see.

Switching to the Behroozi relation did not solve the problem, which means I think the
problem was with one of the occupation components, not the smhm. However,
I can't be sure without a blackbox test of the smhm, so I emailed Peter and requested
he provide me with a tabulation of his 2010 smhm at z=0. I also emailed Alexie
and requested that she provide me with a tabulation of one of her Ncen(M) and Nsat(M),
which I will also use as a blackbox test. In general, the Leauthaud11 model is very
poorly tested, and the Behroozi relation is entirely untested, so these tabulations
will be the blackbox results I use to develop these models.

Note that Alexie assumed h=0.72, whereas Peter uses h=0.7, while the halo catalogs
all assume h=1. These models are also based on M200b, whereas my current catalogs
only have Mvir (quite annoyingly).