Fixes some unintentionally hard-coded choices made in how HeavisideAssembias is implemented.
The basic thing driving this commit is that the lower and upper bounds of the function
being decorated had been hard-coded to be set by "self._lower_bound" and "self._upper_bound".
This works fine if assembly bias is restricted to occupation statistics, but what about
models where this attribute is not present, such as quenching models?

Here is a summary of the changes made to address this problem:

1. ``lower_assembias_bound`` and ``upper_assembias_bound`` are now required keyword arguments
passed to the HeavisideAssembias constructor

2. HeavisideAssembias now uses abstractly-set attributes to pass around the bounds:
lower_bound_key = 'lower_bound_' + self._methodname_to_decorate + '_' + self.gal_type

3. OccupationComponent instances have self._upper_bound ==> self._upper_occupation_bound,
and self._lower_bound ==> self._lower_occupation_bound

4. Appropriate changes are made in the testing suite.