Resolved extremely subtle bug with the cache manager.
If the download manager was initialized, and then the HaloTableCache
was initialized, and then cache manipulations took place that removed
an entry from the ASCII cache log, and then the file that was removed
was downloaded with the previously initialized DownloadManager,
then the newly (re-)downloaded file would not be added to the ascii file.
The reason is that prior to this commit, the way the add_entry_to_cache_log
method worked was to first check self.log for the entry,
and then take no action if it was detected. Since the DownloadManager
was initialized prior to the cache manipulations,
dman.halo_table_cache.log still had the entry that was deleted from disk,
so when dman downloaded the replacement file and called the
self.halo_table_cache.add_entry_to_cache_log method, nothing was added.
This was leading to a mysterious test failure that took many hours to resolve.

Either one of the following two solutions resolves the bug. Both have now been
implemented.

1. Whenever one of the download methods of dman is called, the very first thing
that is done is to call self.halo_table_cache.update_log_from_current_ascii.

2. The add_entry_to_cache_log method now works differently. Now it is no longer
true that no action is taken if no the existing entry is found. Now the passed
entry is added no matter what, and list(set(self.log)) cleans any potential
duplication.