Discovered bug while completing IPython Notebook companion to tutorial 5. In the test suite version of test_tutorial5, the following line is missing from the Size constructor:
self.list_of_haloprops_needed = ['halo_spin']

The test suite still passes when populating a mock
with this model even though this line is omitted. However,
it should *not* pass, because one of the physics functions
requires 'halo_spin'. This is perplexing.
When I run the exact same code within the global environment
of my machine, an exception *is* raised. So there must have been some
update on *master* that did not propagate into this branch yet.