This commit resolves a bug pointed out by @surhud. astropy.cosmology.critical_density function is in physical units, but the rest of the delta_sigma calculation is in comoving units. This commit resolves this by rescaling astropy.cosmology.critical_density by 1/a^3.
There are additional changes to the delta_sigma docstring to clarify the use of comoving units.

There is also a minor change to halotools.sim_manager.FakeSim so that the randomly generated halo catalog has a cosmology attribute. This is helpful to demonstrate the typical call to delta_sigma, which should use the cosmology of the halo catalog from which the galaxy population was generated.