Clarified the installation documentation note on compiler flags.
I don't want to have to remember to change the installation documentation
in the event that in some future commit
I make a change to the choice for the default value of the
compiler optimization flags. So I reworded the note on this installation
topic to make the language apply regardless of how the source code
sets the values of ``extra_compiler_args`` in the ``setup_package.py`` files.