Add support for a ``redshift`` keyword argument to the HodModelFactory constructor.
This feature enables the user to create a custom HOD model in which none of the component models have a redshift defined, but the user desires to define a redshift for the composite model via a keyword argument passed to the HodModelFactory constructor.

The use-case this commit supports is when a user wants to build a composite model for a specific redshift when none of the component models instances have a redshift attribute.
Prior to this commit, the user had two other choices:
1. Overwrite the redshift attribute of the composite model instance returned by the HodModelFactory (without this overwrite, the factory returns an instance with sim_defaults.default_redshift set for the redshift attribute)
2. Overwrite the redshift attribute of one or more component model instances before passing the component model instances to the HodModelFactory constructor.

This commit allows the user to build an instance without ever having to manually set the value of the redshift attribute.