Change prim_galgrop_max value in the test
There are 2 tests in this unit test. After the sign correction, the first test passed, and by lowering the prim_galgrop_max value the second test should pass as well. The reason why it did not pass because the previous chosen value 10^11 is too bright for the default model, which is almost identical to the infinity brightness.